requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
Django==5.1.7
django-admin==2.0.2
django-cors-headers==4.7.0
django-excel-response2==3.0.6
django-six==1.0.5
excel-base==1.0.4
isoweek==1.3.3
pillow==11.1.0
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
screen==1.0.1
six==1.17.0
sqlparse==0.5.3
TimeConvert==3.0.13
tzlocal==5.3.1
xlwt==1.3.0
